# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Freitas Outlet API
  description: API do e-commerce Freitas Outlet — streetwear, luxo & performance
  version: 0.1.0
  contact:
    name: Freitas Outlet
servers:
  - url: http://localhost:3000
    description: Desenvolvimento
  - url: https://freitasoutlet.com.br
    description: Produção
paths:
...
  /api/checkout:
    post:
      summary: Processar checkout
      tags: [Checkout]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action: { type: string, enum: [validate, process] }
                data: { type: object }
                paymentMethod: { type: string, enum: [pix, credit, debit] }
                paymentData: { type: object }
                items: { type: array }
      responses:
        '200': { description: Pedido criado }
        '403': { description: Rejeitado pela análise de fraude }
        '409': { description: Estoque insuficiente }
...
  /api/cliente/perfil:
    put:
      summary: Atualizar perfil
      tags: [Cliente]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                email: { type: string }
                phone: { type: string }
                currentPassword: { type: string }
                newPassword: { type: string }
      responses:
        '200': { description: Perfil atualizado }
...
  /api/webhook:
    post:
      summary: Webhook de pagamento
      tags: [Utils]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type: { type: string }
                data: { type: object }
      responses:
        '200': { description: Recebido }
  /api/contato:
    post:
      summary: Formulário de contato
      tags: [Utils]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                email: { type: string }
                phone: { type: string }
                message: { type: string }
      responses:
        '200': { description: Mensagem enviada }
  /api/verify-email:
    post:
      summary: Verificar e-mail
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string }
      responses:
        '200': { description: E-mail verificado }